Builders: split on continuity violations
RR builders have traditionally failed with an exception when path
continuity is violated. I hesitated before making this change because it
kicks the can down the road. Sure, now new users get a usable return
value instead of an exception, but they'll still be asking why the bot
comes to a stop in the middle.

I was convinced by the realization that the new splitting API is more
general than the previous one and it provides a better foundation for
trajectory sequences (or whatever I decide to call their replacement).