Rename geometry types
Some questions about twists got me thinking about whether twists are
really elements of the tangent space or the Lie algebra. The literature
doesn't seem to follow a single convention, and a few libraries even
dispatch with "twist" altogether for this reason (and to avoid connections
to screw theory).

To match Solà, Deray, and Atchuthan as well as WPILib, I've decided to rename
the old `TwistIncr2d` to `Twist2d`, and the old `Twist2d` to `PoseVelocity2d`.

I also renamed the `trans` and `rot` fields of `Pose2d` to `position` and
`heading`, respectively. The old names were appropriate for `Transform2d`, but
they make less sense now that we use `Pose2d`.

And then there are a few other self-explanatory naming changes.