Feedforward: eliminate actual pose dep
I attempted to improve upon the usual feedforward formulation
by using the actual pose instead of the target pose to
transform the feedforward command into the body frame. This may
be a helpful change for situations where the actual rotation is
decently wrong, though the dependency of feedforward on actual
pose makes tuning/calibration more difficult.